added command line arg processing for console app, made the (quit) function now works more gracfully
git-svn-id: https://commonmusic.svn.sourceforge.net/svnroot/commonmusic/trunk@1705 d60aafaf-7936-0410-ba4d-d307febf7868

+++ src/CmSupport.h
@@ -42,8 +42,8 @@ double cm_decimals(double val, int places); //places=3
 double cm_quantize(double val, double step);
 //tempo=60 beat=.25
 double cm_rhythm_to_seconds(double beats, double tempo, double beat);
-double cm_cents_to_scaler(int cents);
-int cm_scaler_to_cents(double scaler);
+double cm_cents_to_scaler(double cents);
+double cm_scaler_to_cents(double scaler);
 double cm_scaler_to_steps(double scaler);
 double cm_keynum_to_hertz (double kn);
 int cm_keynum_to_pc (double kn);
